+++ jnosql-lite/mapping-lite-column-test/src/test/java/org/eclipse/jnosql/lite/mapping/entities/column/ColumnEntityConverterTest.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.mapping.lite.column;
+package org.eclipse.jnosql.lite.mapping.entities.column;
 
 import jakarta.inject.Inject;
 import org.eclipse.jnosql.communication.TypeReference;
