+++ jnosql-lite/mapping-lite-core-test/src/test/java/org/eclipse/jnosql/lite/mapping/entities/UserTest.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2022 Otávio Santana and others
+ *  Copyright (c) 2020 Otávio Santana and others
  *   All rights reserved. This program and the accompanying materials
  *   are made available under the terms of the Eclipse Public License v1.0
  *   and Apache License v2.0 which accompanies this distribution.
@@ -12,16 +12,13 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.mapping.lite.inheritance;
+package org.eclipse.jnosql.lite.mapping.entities;
 
-import org.eclipse.jnosql.mapping.entities.inheritance.Notification;
-import org.eclipse.jnosql.mapping.entities.inheritance.SocialMediaNotification;
-import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.lite.mapping.metadata.LiteEntitiesMetadata;
+import org.eclipse.jnosql.mapping.entities.User;
+import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.mapping.metadata.EntityMetadata;
 import org.eclipse.jnosql.mapping.metadata.FieldMetadata;
-import org.eclipse.jnosql.mapping.metadata.InheritanceMetadata;
-import org.eclipse.jnosql.mapping.DiscriminatorColumn;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -30,8 +27,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-public class SocialMediaNotificationTest {
+import static org.assertj.core.api.Assertions.assertThat;
 
+public class UserTest {
 
     private EntitiesMetadata mappings;
 
@@ -40,40 +38,44 @@ public class SocialMediaNotificationTest {
     @BeforeEach
     public void setUp() {
         this.mappings = new LiteEntitiesMetadata();
-        this.entityMetadata = this.mappings.get(SocialMediaNotification.class);
+        this.entityMetadata = this.mappings.get(User.class);
     }
 
     @Test
     public void shouldGetName() {
-        Assertions.assertEquals("Notification", entityMetadata.name());
+        Assertions.assertEquals("User", entityMetadata.name());
     }
 
     @Test
     public void shouldGetSimpleName() {
-        Assertions.assertEquals(SocialMediaNotification.class.getSimpleName(), entityMetadata.simpleName());
+        Assertions.assertEquals(User.class.getSimpleName(), entityMetadata.simpleName());
     }
 
     @Test
     public void shouldGetClassName() {
-        Assertions.assertEquals(SocialMediaNotification.class.getName(), entityMetadata.className());
+        Assertions.assertEquals(User.class.getName(), entityMetadata.className());
     }
 
     @Test
     public void shouldGetClassInstance() {
-        Assertions.assertEquals(SocialMediaNotification.class, entityMetadata.type());
+        Assertions.assertEquals(User.class, entityMetadata.type());
     }
 
     @Test
     public void shouldGetId() {
         Optional<FieldMetadata> id = this.entityMetadata.id();
         Assertions.assertTrue(id.isPresent());
+        FieldMetadata fieldMetadata = id.orElseThrow();
+        assertThat(fieldMetadata)
+                .isInstanceOf(FieldMetadata.class);
+
     }
 
     @Test
     public void shouldCreateNewInstance() {
-        SocialMediaNotification notification = entityMetadata.newInstance();
-        Assertions.assertNotNull(notification);
-        Assertions.assertTrue(notification instanceof SocialMediaNotification);
+        User user = entityMetadata.newInstance();
+        assertThat(user)
+                .isNotNull().isInstanceOf(User.class);
     }
 
     @Test
@@ -82,25 +84,16 @@ public class SocialMediaNotificationTest {
         Assertions.assertEquals(4, fields.size());
         Assertions.assertTrue(fields.contains("id"));
         Assertions.assertTrue(fields.contains("name"));
-        Assertions.assertTrue(fields.contains("nickname"));
     }
 
     @Test
     public void shouldGetFieldsGroupByName() {
         Map<String, FieldMetadata> groupByName = this.entityMetadata.fieldsGroupByName();
-        Assertions.assertNotNull(groupByName);
-        Assertions.assertNotNull(groupByName.get("_id"));
-        Assertions.assertNotNull(groupByName.get("name"));
+        assertThat(groupByName).hasSize(4)
+                .containsKeys("_id", "name", "age", "phones");
     }
 
-    @Test
-    public void shouldGetInheritanceMetadata() {
-        InheritanceMetadata inheritance = this.entityMetadata.inheritance()
-                .orElseThrow();
-        Assertions.assertEquals("SocialMediaNotification", inheritance.discriminatorValue());
-        Assertions.assertEquals(DiscriminatorColumn.DEFAULT_DISCRIMINATOR_COLUMN, inheritance.discriminatorColumn());
-        Assertions.assertEquals(SocialMediaNotification.class, inheritance.entity());
-        Assertions.assertEquals(Notification.class, inheritance.parent());
-    }
+
+
 
 }
